luxe/pyproject.toml
[project]
name = "luxe"
version = "0.1.0"
description = "Local coding LLM evaluation + optimization harness for Apple Silicon"
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.8",
    "pyyaml>=6.0",
    "rich>=13.7",
    "typer>=0.12",
    "datasets>=3.0",
    "huggingface-hub>=0.26",
    "psutil>=6.0",
    "jinja2>=3.1",
    "tenacity>=9.0",
    "tqdm>=4.66",
    "ddgs>=9.0",
    "trafilatura>=1.12",
    "prompt-toolkit>=3.0",
]

[project.optional-dependencies]
mlx = [
    "mlx>=0.21",
    "mlx-lm>=0.20",
]
evalplus = [
    "evalplus>=0.3.1",
]
browser = [
    # CDP client for the browser tool. Pure-Python, sync API,
    # ~600 LOC. Optional because not every install needs JS-rendered
    # web access — luxe's static `fetch_url` covers the common case.
    # Headless Chrome itself must be installed separately
    # (`brew install --cask google-chrome`).
    "pychrome>=0.2.4",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
    "mypy>=1.11",
    "bandit>=1.7",
    "pip-audit>=2.7",
    # semgrep is the heaviest dev dep (~100 MB pulled + ~300 MB of
    # default rulesets resolved lazily). Kept in `dev` (not a
    # separate extra) because /review's `security_taint` tool degrades
    # gracefully when it's absent — users who skip it still get
    # `security_scan` (bandit) + `deps_audit` (pip-audit).
    "semgrep>=1.86",
]
# `bfcl-eval` is intentionally NOT listed here.
# It pins tree-sitter==0.21.3, which conflicts with evalplus (>=0.22). Install
# BFCL into its own venv when you want to run that benchmark:
#
#   uv venv .venv-bfcl
#   .venv-bfcl/bin/pip install bfcl-eval httpx rich typer tenacity psutil pyyaml pydantic
#   .venv-bfcl/bin/python scripts/run_phase_a.py --candidate <id> --bench bfcl_v3

[project.scripts]
lux = "harness.cli:app"
luxe = "luxe_cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["harness", "benchmarks", "personal_eval", "luxe_cli", "shared"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
